requirements.txt
autopep8==1.3.4
certifi==2018.1.18
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.2
django-cors-headers==2.1.0
django-debug-toolbar==1.9.1
django-filter==1.1.0
django-js-asset==1.0.0
django-mptt==0.9.0
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-silk==2.0.0
django-viewflow==1.2.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
Flask==0.12.2s
gprof2dot==2016.10.13
idna==2.6
imagekit==1.0.6
itsdangerous==0.24
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mod-wsgi==4.5.22
olefile==0.44
openapi-codec==1.3.2
Pillow==4.3.0
psycopg2==2.7.4
pyasn1==0.4.2
pycodestyle==2.3.1
Pygments==2.2.0
PyJWT==1.5.3
PyMySQL==0.7.11
python-dateutil==2.6.1
pytz==2018.3
requests==2.18.4
rsa==3.4.2
simplejson==3.13.2
six==1.11.0
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.14.1
